A very simple CI script
Using the system python3 at the moment (the `python-version` specified
in the build matrix is ignore -- though it happens to currently be the
same as system py3 on ubuntu-latest), because any python installed by
`setup-python` seg faults on CDLL calls.

I imagine there's *something* wrong with my code that exposes this
behavior, but what that might be remains a mystery to me.